Improved playlist traversal.
* Next/Prev buttons still not working
As it is designed now, the playlist is an array, and the playlist items
hold their index in the element. This works until elements are deleted.
I might select another method that assigns a unique ID to its element,
and then using the ID as the playlist key, instead of an ordered array.